Refactor: ClientSettingsPacket
